Conflicts:
	platform/android/MapboxGLAndroidSDK/proguard-rules.pro
	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/maps/MapView.java
	platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/maps/widgets/MyLocationView.java
	src/mbgl/style/layers/line_layer.cpp

+++ platform/android/src/style/android_conversion.hpp
@@ -80,7 +80,8 @@ inline optional<Value> toValue(const mbgl::android::Value& value) {
     } else if (value.isString()) {
         return { value.toString() };
     } else if (value.isNumber()) {
-       return { value.toNumber() };
+        //Need to cast to a double here as the float is otherwise considered a bool...
+       return { (double) value.toNumber() };
     } else {
         return {};
     }
